Added ability to add section to binary.
Added ability to add section to binary. This is accessible via image_t::add_section, and will append a section to the binary and return a buffer containing the whole binary (including the new section). The section is not applied to the runtime image, only returned in a buffer from the function.